+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added support for INTEGER columns in lookup tables.
 - Added support for provided user-defined dictionaries in the configuration.
+- Now displaying errors at the end of the logs to make them more obvious.
 
 **Version 1.0**