+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Version 1.0**
 
+- Added support for all-spaces regular expressions.
 - Now displaying errors at the end of the logs to make them more obvious.
 - Added support for provided user-defined dictionaries in the configuration.
 - Added support for INTEGER columns in lookup tables.