+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Validation Translation Version History
 
+**Version 1.1**
+
+- Updated validation library from version 2.25 to version 2.27.
+- Updated naaccr-xml library from version 8.9 to version 8.10.
+- Updated cs-staging library from version 02.05.50.9 to version 02.05.50.10.
+- Updated log4j-api from version 2.17.0 to version 2.17.2.
+
 **Version 1.0**
 
 - Added support for all-spaces regular expressions.